e1.11.0 Configuration refactored and simplified (interface changed).

+++ storage.go
@@ -153,7 +153,7 @@ func (l *Logger) getDiskFreeSpace(path string) (int64, error) {
 	if err := syscall.Statfs(path, &stat); err != nil {
 		return 0, fmtErrorf("failed to get disk stats for '%s': %w", path, err)
 	}
-	availableBytes := int64(stat.Bavail) * int64(stat.Bsize)
+	availableBytes := int64(stat.Bavail) * stat.Bsize
 	return availableBytes, nil
 }
